put qt ffi behind a feature flag
