+++ CMakeLists.txt
@@ -41,7 +41,7 @@ else()
 endif()
 
 add_cargo(MANIFEST Cargo.toml)
-add_crate(gui LIBRARY)
+add_crate(gui LIBRARY ARGS --features "qt_ffi")
 add_crate(obkrnl
     TOOLCHAIN ${kernel_toolchain}
     VENDOR "unknown"
